submission/source_code/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sira"
version = "0.1.0"
description = "Service-oriented SIR vector field training and inference pipeline"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116,<1.0",
    "httpx>=0.28,<1.0",
    "matplotlib>=3.9,<4.0",
    "numpy>=2.2,<3.0",
    "pandas>=2.2,<3.0",
    "PyYAML>=6.0,<7.0",
    "scipy>=1.15,<2.0",
    "torch>=2.6,<3.0",
    "tqdm>=4.67,<5.0",
    "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "ruff>=0.11,<0.12",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["sira*", "src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = ["B008"]